# api-spec.yaml
openapi: 3.0.3

info:
  title: Simple Book API
  version: 1.0.0
  description: Basic CRUD API for books (in-memory storage)

servers:
  - url: http://localhost:3000

paths:

  /books:
    get:
      summary: Get all books (with optional filters)
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: author
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of books

    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: Book created

  /books/{id}:
    get:
      summary: Get book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Single book
        '404':
          description: Book not found

    put:
      summary: Replace entire book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '200':
          description: Book updated

    delete:
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Book deleted

components:

  schemas:

    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        year:
          type: integer
        rating:
          type: number
        available:
          type: boolean

    BookInput:
      type: object
      required: [title, author]
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        year:
          type: integer
        rating:
          type: number
        available:
          type: boolean

    BookPatch:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        year:
          type: integer
        rating:
          type: number
        available:
          type: boolean
          